+++ examples/rag-chat-api/api_config.ini
@@ -9,3 +9,7 @@ postgres_user=
 auth_token=
 llm_host=
 llm_model=meta-llama/Llama-2-13b-chat-hf
+
+[tables]
+embeddings_table=
+data_table=
